// src/lib/models/StatsResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type StatsResponse = {
    totalMinutesFocused?: number;
    totalCompletedSessions?: number;
    totalAbandonedSessions?: number;
    completionRate?: number;
    currentStreak?: number;
    totalPoints?: number;
    level?: number;
    levelTitle?: string;
    totalTasksDone?: number;
};